requirements.txt
python-chess
pygame
streamlit